change float16 to flost32 to prevent overflow

+++ basenji/dataset.py
@@ -111,7 +111,7 @@ class SeqDataset:
         sequence = tf.cast(sequence, tf.float32)
 
       # decode targets
-      targets = tf.io.decode_raw(parsed_features[TFR_OUTPUT], tf.float16)
+      targets = tf.io.decode_raw(parsed_features[TFR_OUTPUT], tf.float32)
       if not raw:
         targets = tf.reshape(targets, [self.target_length, self.num_targets])
         targets = tf.cast(targets, tf.float32)
